requirements_.txt
absl-py==2.2.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
ecdsa==0.19.1
fastapi==0.110.0
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
groq==0.25.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
keras==3.10.0
lazy_loader==0.4
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.9
networkx==3.4.2
numpy==1.24.3
opencv-python-headless==4.9.0.80
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pillow==10.2.0
protobuf==4.25.7
pyasn1==0.6.1
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.19.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.7
requests==2.32.3
rich==14.0.0
rsa==4.9.1
scikit-image==0.22.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
starlette==0.36.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tifffile==2025.5.21
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.27.1
Werkzeug==3.1.3
wrapt==1.17.2
